move IPS/header handling out of snesreader
did you know that bsnes v073 actually supported IPS softpatching the
entire time? and that it was implemented in a completely different place
than UPS patching was? and that it also wasn't documented anywhere or
presented to the user at all, and that it was always carried out
regardless of whether or not the user had disabled soft patching in the
file browser? for fuck's sake, byuu