Create breakpoints using the command line arguments
This allows me to use a script to automatically fill the bsnes debugger
with breakpoints. For instance, the following bash line:

    bsnes --show-debugger `grep BREAKPOINT "$symfile" | awk '{print "-b " $1 ":x"}'` "$romfile"

Will open the debugger window and create a CPUBus execute breakpoint
for all symbols containing BREAKPOINT in the symbol file.

See `bsnes --help` or `Application::printArguments` for the format of
the breakpoint string.