Allow the user to sort the Sprite Viewer columns
Decided to move the OAM name table bit into the "Char" column so it will
be included in the sort.

Also decided that refresh() should not change the user selected row.
This required me remember the selected row so it could be re-selected
after QTreeWidget has sorted the model. Unfortunately this can lead to
the occasional infinite signal call loop, necessitating the
inRefreshCall test to ensure refresh() is only called once.